+++ lib/gds_api/test_helpers/email_alert_api.rb
@@ -408,6 +408,53 @@ module GdsApi
           .to_return(status: 404)
       end
 
+      def stub_email_alert_api_bulk_unsubscribe(subscriber_list_id:)
+        stub_request(:post, "#{EMAIL_ALERT_API_ENDPOINT}/subscriber-lists/#{subscriber_list_id}/bulk-unsubscribe")
+          .to_return(status: 202)
+      end
+
+      def stub_email_alert_api_bulk_unsubscribe_with_message(subscriber_list_id:, body:, sender_message_id:)
+        stub_request(:post, "#{EMAIL_ALERT_API_ENDPOINT}/subscriber-lists/#{subscriber_list_id}/bulk-unsubscribe")
+        .with(
+          body: {
+            body: body,
+            sender_message_id: sender_message_id,
+          }.to_json,
+        ).to_return(status: 202)
+      end
+
+      def stub_email_alert_api_bulk_unsubscribe_not_found(subscriber_list_id:)
+        stub_request(:post, "#{EMAIL_ALERT_API_ENDPOINT}/subscriber-lists/#{subscriber_list_id}/bulk-unsubscribe")
+          .to_return(status: 404)
+      end
+
+      def stub_email_alert_api_bulk_unsubscribe_not_found_with_message(subscriber_list_id:, body:, sender_message_id:)
+        stub_request(:post, "#{EMAIL_ALERT_API_ENDPOINT}/subscriber-lists/#{subscriber_list_id}/bulk-unsubscribe")
+        .with(
+          body: {
+            body: body,
+            sender_message_id: sender_message_id,
+          }.to_json,
+        ).to_return(status: 404)
+      end
+
+      def stub_email_alert_api_bulk_unsubscribe_conflict_with_message(subscriber_list_id:, body:, sender_message_id:)
+        stub_request(:post, "#{EMAIL_ALERT_API_ENDPOINT}/subscriber-lists/#{subscriber_list_id}/bulk-unsubscribe")
+        .with(
+          body: {
+            body: body,
+            sender_message_id: sender_message_id,
+          }.to_json,
+        ).to_return(status: 409)
+      end
+
+      def stub_email_alert_api_bulk_unsubscribe_bad_request(subscriber_list_id:, body:)
+        stub_request(:post, "#{EMAIL_ALERT_API_ENDPOINT}/subscriber-lists/#{subscriber_list_id}/bulk-unsubscribe")
+        .with(
+          body: { body: body }.to_json,
+        ).to_return(status: 422)
+      end
+
     private
 
       def get_subscriber_response(id, address, govuk_account_id)